pyproject.toml
[tool.poetry]
name = "syspro-connect"
version = "0.1.0"
description = ""
authors = ["w-ejt <97180065+w-ejt@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "syspro_connect", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^1.5.3"
sqlalchemy = "^1.4.46"
pyodbc = "^4.0.35"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
